+++ src/components/Transition.vue
@@ -21,16 +21,22 @@ export default {
     position: { type: String, required: true }
   },
   methods: {
+    // todo - consider will-change
     leave(el) {
+      if (
+        this.$parent.singular ||
+        (this.$parent.oneType && this.$parent.$el.childNodes.length === 1)
+      ) {
+        return;
+      }
       const position = this.position.split("-");
       // https://forum.vuejs.org/t/transition-group-move-class-not-occuring-in-the-array/6381/5
       // these rules ensure the toast stays where it is
       const { height, width, marginBottom } = window.getComputedStyle(el);
       // when the last toast removed the container collapses hence the need for the width subtraction
       el.style.left =
-        (el.offsetLeft -
-          (el.parentNode.childNodes.length === 1 ? parseInt(width) : 0)) /
-          2 +
+        el.offsetLeft -
+        (el.parentNode.childNodes.length === 1 ? parseInt(width) : 0) +
         "px";
       el.style.top = el.offsetTop + "px";
       if (position[0] === "center") {
@@ -51,12 +57,25 @@ export default {
       el.style.width = width;
       el.style.position = "absolute";
     },
-    beforeEnter() {
+    // eslint-disable-next-line no-unused-vars
+    beforeEnter(el) {
       // no delay on making space for notification
       this.$el.childNodes.forEach(node => delete node.dataset.delayed);
+      if (el.__vue__.status.delayed) {
+        el.dataset.delayed = true;
+        el.classList.add("vt-move");
+        delete el.__vue__.status.delayed;
+      }
+
+      // if (el.__vue__.status.delayed) {
+      //   el.dataset.delayed = true;
+      //   el.classList.add("vt-move");
+      //   delete el.__vue__.status.delayed;
+      // }
     },
     afterEnter(el) {
       el.classList.add("vt-default-position");
+      el.removeAttribute("data-delayed");
     },
     beforeLeave(el) {
       // this ensures that notifications won't move until the other has been removed